Moved raw structure into main package, but as a private member.
Shuffled things around as a result.
